오류해결을 위한 코드수정 : dc 본문을 가져오지 못하는 오류
- response = requests.get(url, headers=header) 가 정상적으로 작동하는데(response.status_code==200) 빈 값을 불러온다
- 충분한 시간을 주어서 해결하고자 한다 : time.sleep(2)
