+++ crawling/src/dcinside_url_crawler.py
@@ -1,11 +1,11 @@
 # Made by Hansol Lee 20230925
 
-from crawling_tool import get_gall_id, get_url_base, get_max_num, get_last_page
+from crawling_tool import get_gall_id, get_url_base, get_max_num, get_last_page, get_driver
 from bs4 import BeautifulSoup
 import utility_module as util
 import requests
 import pandas as pd
-import re
+
 
 # keyword = "토스"
 # 2015-02-26 ~
