+++ crawling/tests/test_dcinside.py
@@ -17,14 +17,15 @@ gall_url = {
     "토스": "https://gall.dcinside.com/mgallery/board/lists/?id=toss"
 }
 
-# 해외주식, 코스피, 체크카드, SFF, 토스
+#
 # [1. url 크롤링]
 # get_url_dcinside(gall_url["토스"], keyword)
 
 # get_url_dc_toss(keyword, gall_url["토스"])
 
+# 편의점
 # [2. content 크롤링]
-get_content_dc(gall_url["해외주식"])
+get_content_dc(gall_url["편의점"])
 
 #########################################
 # url = "https://gall.dcinside.com/board/lists/?id=skwyverns_new1&s_type=search_subject_memo&s_keyword=.E3.85.87.E3.85.87"
